+++ src/main/java/com/windstorm/management/service/cell/CellService.java
@@ -2,8 +2,8 @@ package com.windstorm.management.service.cell;
 
 import org.springframework.stereotype.Service;
 
-import com.windstorm.management.controller.cell.request.CellAddMember;
-import com.windstorm.management.controller.cell.request.CellCreate;
+import com.windstorm.management.controller.admin.cell.request.CellAddMember;
+import com.windstorm.management.controller.admin.cell.request.CellCreate;
 import com.windstorm.management.controller.cell.response.CellResponse;
 import com.windstorm.management.domain.cell.Cell;
 import com.windstorm.management.implement.cell.CellAppender;
